+++ lib/instream/series.ex
@@ -8,8 +8,8 @@ defmodule Instream.Series do
         use Instream.Series
 
         series do
-          database    :my_database
-          measurement :cpu_load
+          database    "my_database"
+          measurement "cpu_load"
 
           tag :host, default: "www"
           tag :core
@@ -113,9 +113,21 @@ defmodule Instream.Series do
   @doc """
   Defines the database for the series.
   """
-  defmacro database(name) do
-    name = to_string(name)
+  defmacro database(name) when is_atom(name) do
+    quote do
+      IO.write :stderr, """
+      The series "#{ __MODULE__ }"
+      has a database configured using an atom.
 
+      This has been deprecated in favor of strings.
+      Configuration using atoms will eventually be removed and stop working.
+      """
+
+      database to_string(unquote(name))
+    end
+  end
+
+  defmacro database(name) do
     quote do
       unquote(__MODULE__).__attribute__(__MODULE__, :database, unquote(name))
     end
@@ -136,9 +148,21 @@ defmodule Instream.Series do
   @doc """
   Defines the measurement of the series.
   """
-  defmacro measurement(name) do
-    name = to_string(name)
+  defmacro measurement(name) when is_atom(name) do
+    quote do
+      IO.write :stderr, """
+      The series "#{ __MODULE__ }"
+      has a measurement configured using an atom.
 
+      This has been deprecated in favor of strings.
+      Configuration using atoms will eventually be removed and stop working.
+      """
+
+      measurement to_string(unquote(name))
+    end
+  end
+
+  defmacro measurement(name) do
     quote do
       unquote(__MODULE__).__attribute__(__MODULE__, :measurement, unquote(name))
     end