+++ test/instream/connection_test.exs
@@ -17,8 +17,8 @@ defmodule Instream.ConnectionTest do
     use Instream.Series
 
     series do
-      database    :test_database
-      measurement :data_write_struct
+      database    "test_database"
+      measurement "data_write_struct"
 
       tag :bar
       tag :foo